Fixed the last bugs. Only bug left is that we can't split space into decimals so when points are to close we have infinite loop

+++ quadtree.hpp
@@ -15,10 +15,10 @@ using namespace std;
 
 struct DimensionSpace
 {
-    int start;
-    int end;
+    float start;
+    float end;
 public:
-    DimensionSpace(int start, int end);
+    DimensionSpace(float start, float end);
     DimensionSpace() {};
 };
